useGame hook
